VIC-8414: Adjust battery filter in calm mode (#3920)
* Update sampling period for battery voltage low pass filter based on calm mode state since RobotState send rate differs

* Defer reseting the battery voltage filter until after the battery has reconnected. Otherwise, the voltage could stay at a lower-than-actual value for a little longer than it needs to. e.g. Robot is placed on a charger while the battery is hot. It would never actually reset the voltage because the battery is disconnected at the time the onCharger state changed and it won't reset later when the battery reconnects after cooldown.

* Added unit test for new LowPassFilter functions

* Fixed RobotState.status masking to be against u32. Doesn't fix any bugs, but might fix future ones if we ever move those enums around.

+++ animProcess/src/cozmoAnim/micData/micDataProcessor.cpp
@@ -534,8 +534,8 @@ MicDirectionData MicDataProcessor::ProcessMicrophonesSE(const AudioUtil::AudioSa
   PolicySetAbsoluteOrientation(robotAngle);
   // Note that currently we are only monitoring the moving flag. We _could_ also discard mic data when the robot
   // is picked up, but that is being evaluated with design before implementation, see VIC-1219
-  const bool robotIsMoving = static_cast<bool>(robotStatus & (uint16_t)RobotStatusFlag::IS_MOVING);
-  const bool isLowPowerMode = static_cast<bool>(robotStatus & (uint16_t)RobotStatusFlag::CALM_POWER_MODE);
+  const bool robotIsMoving = static_cast<bool>(robotStatus & (uint32_t)RobotStatusFlag::IS_MOVING);
+  const bool isLowPowerMode = static_cast<bool>(robotStatus & (uint32_t)RobotStatusFlag::CALM_POWER_MODE);
   const bool robotStoppedMoving = !robotIsMoving && _robotWasMoving;
   _robotWasMoving = robotIsMoving;
